feat(插件化): 迁移editro功能
