+++ src/core/ServersPlugin.ts
@@ -2,7 +2,7 @@
  * @Author: 秦少卫
  * @Date: 2023-06-20 12:52:09
  * @LastEditors: 秦少卫
- * @LastEditTime: 2023-06-27 23:13:22
+ * @LastEditTime: 2023-06-28 12:30:14
  * @Description: 内部插件
  */
 
@@ -17,12 +17,11 @@ class ServersPlugin {
   public canvas: fabric.Canvas;
   public editor: IEditor;
   static pluginName = 'ServersPlugin';
-  static apis = ['insert', 'insertSvgFile'];
+  static apis = ['insert', 'insertSvgFile', 'getJson', 'dragAddItem'];
   // public hotkeys: string[] = ['left', 'right', 'down', 'up'];
   constructor(canvas: fabric.Canvas, editor: IEditor) {
     this.canvas = canvas;
     this.editor = editor;
-    // this.init();
   }
 
   insert() {
@@ -45,6 +44,30 @@ class ServersPlugin {
         });
       });
     });
+  }
+
+  getJson() {
+    return this.canvas.toJSON(['id', 'gradientAngle', 'selectable', 'hasControls']);
+  }
+
+  /**
+   * @description: 拖拽添加到画布
+   * @param {Event} event
+   * @param {Object} item
+   */
+  dragAddItem(event: DragEvent, item: fabric.Object) {
+    const { left, top } = this.canvas.getSelectionElement().getBoundingClientRect();
+    if (event.x < left || event.y < top || item.width === undefined) return;
+
+    const point = {
+      x: event.x - left,
+      y: event.y - top,
+    };
+    const pointerVpt = this.canvas.restorePointerVpt(point);
+    item.left = pointerVpt.x - item.width / 2;
+    item.top = pointerVpt.y;
+    this.canvas.add(item);
+    this.canvas.requestRenderAll();
   }
 
   destroy() {
