+++ src/views/home/index.vue
@@ -140,7 +140,7 @@ import layer from '@/components/layer.vue';
 import attribute from '@/components/attribute.vue';
 
 // 右键菜单
-import mouseMenu from '@/components/contextMenu/index.vue';
+// import mouseMenu from '@/components/contextMenu/index.vue';
 
 // 功能组件
 import CanvasEventEmitter from '@/utils/event/notifier';
